+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 7.3.0
+Version 7.3.1
 
 # Changelog
 
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 7.3.1 - 2024-04-16
 
 ### Fixed